Crate ra_assists

ra_assists crate provides a bunch of code assists, also known as code actions (in LSP) or intentions (in IntelliJ).

An assist is a micro-refactoring, which is automatically activated in certain context. For example, if the cursor is over ,, a "swap ," assist becomes available.

Modules

ast_transform
AstTransformers are functions that replace nodes in an AST and can be easily combined.

utils
Assorted functions shared by several assists.

Structs

AssistAction

AssistId
Unique identifier of the assist, should not be shown to the user directly.

AssistLabel

GroupLabel

ResolvedAssist

Enums

AssistFile

Functions

resolved_assists
Return all the assists applicable at the given position.

unresolved_assists
Return all the assists applicable at the given position.